Honor QPrinter::setFullPage(true) on Windows too (no margins)
fullPage means the print engine shouldn't add any margins,
the application will take care of that. That's already what happens
on Linux and Mac, but the Windows print engine was offset-ting
everything to the bottom right by the value of the margins, erroneously.

As noted in QTBUG-95927, the workaround for this bug was to call
  printer.setPageMargins(QMargins(0,0,0,0));
when using printer.setFullPage(true), and this fix is compatible
with that workaround, existing apps won't be broken.

[ChangeLog][QtPrintSupport][Windows] setFullPage(true) now behaves
as expected, i.e. the QPrinter margins are ignored and the drawing's
(0, 0) is the topleft corner of the page. This is what setFullPage(true)
is documented to do, and how it was already working on other operating
systems. If this causes regressions in your application, consider
removing the call to setFullPage(true) so that the painting honors
the margins again.

Fixes: QTBUG-119003
Fixes: QTBUG-95927
Change-Id: Ia3d621302bf752833002614303dd64128027163a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>