Revert "synqt.cpp: scan for and reject #pragma once"
This breaks qtapplicationmanager, which uses #pragma once exclusively.
That repo is fixed for 6.7/dev, but backporting that huge patch to 6.6
(dead) and 6.5 (LTS) is not very practical and also odd to have that
late in the LTS release cycle.

This reverts commit 656fb759115467de31b771b0138411e73d2b643a.

Change-Id: I04576453a1b06878d2afa2ef1e6dda038ce864ad
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 7576e4339d02442cb0fc148924bce98bdc46c3e0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>