tst_QMutex[qnx]: increase system timer resolution
I don't know what it is but it's not 1, because it is quite
often flaky in these tests. A single failure leads to deadlock
that takes 5 minutes to fail. Then a rerun might do another
5 minutes and fail the integration, or it will pass but take
longer to do so.

Pick-to: 6.6 6.5
Change-Id: I188276df7800b00a20dbe39edee91c582f0a82a7
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 6aee3342e9d0829bb113583c2b6ecc6c94492d2c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>