CMake: Add public helper _qt_internal_add_phony_target
This function can be used to create a custom target like
update_translations or all_qmllint that should be excluded from build
and be triggered by the user.

The function works around certain peculiarities of the Visual Studio
project generators that are described at length in QTBUG-115166 and
associated patches.

Follow-up patches in qttools and qtdeclarative will use this function
and remove duplicated code.

Task-number: QTBUG-115166
Task-number: QTBUG-118980
Change-Id: I87b89a658f35a5a09e3f7b04bdd5cae2166f8d62
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>